fix test_clickhouse.py to pass tests with new operator (0.13+) versions, see https://github.com/Altinity/clickhouse-operator/issues/647
