+++ tests/templates/tpl-clickhouse-19.11.yaml
@@ -12,5 +12,5 @@ spec:
       - name: default
         spec:
           containers:
-            - name: clickhouse
+            - name: clickhouse-pod
               image: yandex/clickhouse-server:19.11.12.69
