+++ tests/templates/tpl-clickhouse-20.8.yaml
@@ -13,4 +13,4 @@ spec:
         spec:
           containers:
             - name: clickhouse-pod
-              image: yandex/clickhouse-server:20.4
+              image: yandex/clickhouse-server:20.8
