+++ tests/templates/tpl-clickhouse-auto.yaml
@@ -14,7 +14,7 @@ spec:
       - name: default
         spec:
           containers:
-            - name: clickhouse
+            - name: clickhouse-pod
               image: yandex/clickhouse-server:20.8.7.15
           #    command:
           #      - "/bin/bash"
